backend/requirements.txt
# -----------------------------
# Core Backend
# -----------------------------
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
pydantic==2.11.3
python-dotenv==1.2.1

# -----------------------------
# Embeddings & Transformers
# -----------------------------

# accelerate==1.12.0
# sentence-transformers==5.1.2
# transformers==4.57.3
# torch==2.9.1

sentence-transformers==2.6.1
transformers==4.41.2
torch==2.2.2
accelerate==0.30.1
# -----------------------------
# LangChain & Text Processing
# -----------------------------

langchain==1.1.0
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-groq==1.1.0
langchain-text-splitters==1.0.0
huggingface_hub[hf_xet]


# -----------------------------
# Vision & Image Processing
# -----------------------------
pillow==12.0.0
easyocr==1.7.2
opencv-python-headless==4.9.0.80


# -----------------------------
# Vector Database
# -----------------------------
pinecone==7.3.0
chromadb==1.3.5

# -----------------------------
# LLM Provider
# -----------------------------
groq==0.30.0

# -----------------------------
# File Handling
# -----------------------------
python-docx==1.2.0
pypdf==6.5.0
docx2txt==0.9

# -----------------------------
# Utilities
# -----------------------------
tqdm
numpy==1.26.4
uuid

# -----------------------------
# Evaluation
# -----------------------------
ragas 
datasets 
openai==2.15.0
tiktoken

# =========================
# Async / web compatibility
# =========================
aiofiles==24.1.0
pi_heif
unstructured
unstructured-client==0.33.0